refactor : 같은 도메인의 컴포넌트 폴더로 묶기
