+++ src/app/(menu)/festival-news/faq/[id]/page.tsx
@@ -1,4 +1,4 @@
-import PostingContainer from '@/components/PostingContainer';
+import PostingContainer from '@/components/Posting/PostingContainer';
 
 export default async function page() {
   return <PostingContainer />;
