+++ src/app/(menu)/festival-news/faq/page.tsx
@@ -1,4 +1,4 @@
-import BoardContainer from '@/components/BoardContainer';
+import BoardContainer from '@/components/Board/BoardContainer';
 
 export default async function page() {
   return <BoardContainer />;
