+++ src/app/(menu)/festival-news/press-release/[id]/page.tsx
@@ -1,4 +1,4 @@
-import PostingContainer from '@/components/PostingContainer';
+import PostingContainer from '@/components/Posting/PostingContainer';
 
 export default function page() {
   return <PostingContainer />;
