+++ src/app/(menu)/festival-news/press-release/page.tsx
@@ -1,4 +1,4 @@
-import BoardContainer from '@/components/BoardContainer';
+import BoardContainer from '@/components/Board/BoardContainer';
 
 export default function page() {
   return <BoardContainer />;
