+++ src/app/admin/main-image/page.tsx
@@ -1,4 +1,4 @@
-import GalleryContainer from '@/components/GalleryContainer';
+import GalleryContainer from '@/components/Gallery/GalleryContainer';
 
 export default function page() {
   return <GalleryContainer />;
